RavenDB-18645 - GET /databases/*/collections/stats + detailed - moved to processor
